requirements.txt
opencv-python
numpy
ultralytics
torch
torchvision
transformers
pillow
ollama
requests
timm
